a lot of things
- drive code
- basic intaking command
- constants
KEEP IN MIND: none of this is final. a lot is very very rough. but i figured better commit now for y'all to sync to and then edit (since we're editing actual code a lot more now) than finish all these things on my own and then mess up everything with a merge