Update #2
Refactored Drivetrain.java, which allowed me to fix and finish refactoring RobotContainer and Drive for now. There were a few remaining errors in Update #1 that had not been addressed.

In particular, Drivetrain now requires a left and right joystick in its constructor so that RobotController does not need to be passed as a parameter to the Drivetrain.