Update #3
Huge update.
The most important thing was adding a bit of additional structure to most of the subsystems so that start/stop/toggle commands could be made to control all subsystems.

Refactored many names, most in minor ways. Also edited or added some comments, edited scope of variables, and order of code to group together similar calls.

Trying not to do much on spacing, since spacing is clearly very inconsistent depending on who contributed to the code, and this is honestly not important at the moment.

Shooter is now Flywheel throughout the code. Both were very prevalent so I decided to go with Flywheel as a consistent name, since eliminating Flywheel in favor of Shooter would have been much harder.