More fixes!
Did not get a chance to test on robot yet, but have tentative fixes for hopper and tower not toggling from commands, as well as some updates to Vijay's code to control tower/flywheel behavior when ramping down from a shot.  Also added a limelight toggle button for the operator. Implemented a previously blank command that reverses only the hopper, although its need is questionable. May have fixed my unjam command for tower (which is bound to a joystick), which reverses both the tower and hopper - more useful than hopper alone!