various fixes to many subsystems
tower and flywheel have good shooting functionality now - not perfect, but mainly flywheel needs tuning PID.

reverse/unjam does work but it's worth noting that the design of the tower makes reversing possible to actually cause a jam

"hold" to climb is now implemented with a 1 sec delay on the driver's controller.

the robot can get stuck in buggy states from time to time so this is a major priority to understand why this happens......logic errors and conflicting commands are likely the cause, maybe something with the scheduler but probably not.