implemented auto routines
Expanded existing auto code to encapsulate the idea of a routine (but did not build a Routine class, seemed unnecessary for now).  Auto routines now just specify a string which is their name in the chooser, rather than specifying a path.  This string name for the auto routine is used in RobotContainer to construct the details for that particular routine.

I also generalized the idea of ShootThree to ShootNTimes so that we can attempt to shoot an arbitrary number of times during autonomous.